+++ configs/nvim-0.11/plugin/40_plugins.lua
@@ -8,11 +8,9 @@
 --
 -- Use this file to install and configure other such plugins.
 
--- Make concise helpers for installing/adding plugins in two stages.
--- Add some plugins now if Neovim is started like `nvim -- some-file` because
--- they are needed during startup to work correctly.
-local add, now, later = MiniDeps.add, MiniDeps.now, MiniDeps.later
-local now_if_args = vim.fn.argc(-1) > 0 and now or later
+-- Make concise helpers for installing/adding plugins in two stages
+local add, later = MiniDeps.add, MiniDeps.later
+local now_if_args = _G.Config.now_if_args
 
 -- Tree-sitter ================================================================