+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2025-10-16
+
+- Move `now_if_args` startup helper to 'init.lua' as `_G.Config.now_if_args` to be directly usable from other config files.
+
 ## 2025-10-13
 
 - Initial release.